Fixes for:
1. pep8 compliance
2. rename tests and fix for recent changes
3. Utility script to extract a revision from a dump file
