+++ tox.ini
@@ -1,0 +1,3 @@
+[pep8]
+max-line-length = 160
+
